1. Fixed server not unbinding port bug.\2. Separated client handling functionality into ClientHandler.py\3. Separated client code into Client folder, server code into Server folder.\ 4. Created Common folder for code that can be shared between client and server.
